[Review] 코드리뷰 받을 주석 추가
