+++ MusicRecordApp/Helper/Date+Extension.swift
@@ -75,7 +75,7 @@ extension Date {
     /// 데이터에 저장하는 쿼리명
     func formattedDate() -> String {
         let dateFormatter = DateFormatter()
-        dateFormatter.dateFormat = "yyyy-MM-dd"
+        dateFormatter.dateFormat = "yyyy년 MM월 dd일"
         return dateFormatter.string(from: self)
     }
