+++ MusicRecordApp/View/Home/Home.swift
@@ -19,6 +19,7 @@ struct Home: View {
     /// Animation Namespace
     @Namespace private var animation
     
+    /// ✍️ Mark: Query, Filter
     @Query private var dateRecord: [Record]
     
     var recordArray: [Record] {
@@ -188,181 +189,4 @@ struct Home: View {
     }
     
 }
-
-struct RecordContentView: View {
-    let imageManager = ImageManager()
-    var songInfo: SongInfo
-    
-    @State private var text: String = ""
-    @State private var showingAlert = false
-    @State private var isPlaying = false
-    @Environment(\.modelContext) var modelContext
-    @EnvironmentObject var sharedDateManager: SharedDataManager
-    @Binding var selectedDate: String
-    
-    @Query private var dateRecord: [Record]
-    
-    var recordArray: [Record] {
-        return dateRecord.filter { record in
-            record.date == selectedDate
-        }
-    }
-    
-    var recordText: String
-    
-    var body: some View {
-        VStack {
-            HStack {
-                Spacer()
-                menuButton()
-            }
-            contentView()
-        }
-        .alert("정말 삭제하시겠습니까?", isPresented: $showingAlert) {
-            Button("아니오", role: .cancel) {}
-            Button("네", role: .destructive) {
-                deleteRecord()
-            }
-        }
-        .padding()
-    }
-    
-    @ViewBuilder
-    func contentView() -> some View {
-        
-        ZStack {
-            
-            AsyncImage(url: URL(string: songInfo.image))
-                .frame(width: 200, height: 200)
-                .cornerRadius(20)
-                .clipped()
-            
-            Button {
-                playToggle()
-            } label: {
-                if songInfo.id == sharedDateManager.nowPlayingID {
-                    Image(systemName: "pause.circle.fill")
-                        .font(.system(size: 50)) // 아이콘 크기 조절
-                        .foregroundColor(.white) // 아이콘 색상
-                        .background(Color.black.opacity(0.4)) // 투명한 검은색 배경
-                        .clipShape(Circle()) // 원형으로 클리핑
-                } else {
-                    Image(systemName: "play.circle.fill")
-                        .font(.system(size: 50)) // 아이콘 크기 조절
-                        .foregroundColor(.white) // 아이콘 색상
-                        .background(Color.black.opacity(0.4)) // 투명한 검은색 배경
-                        .clipShape(Circle()) // 원형으로 클리핑
-                }
-                
-            }
-            
-            
-        }
-        
-        Text(songInfo.title)
-            .font(.title)
-            .fontWeight(.bold)
-            .padding(.top, 8)
-        
-        Text(songInfo.singer)
-            .font(.headline)
-            .foregroundColor(.secondary)
-            .padding(.bottom, 8)
-        
-        Text(recordText)
-            .font(.body) // 글꼴 크기 설정
-            .frame(maxWidth: .infinity, maxHeight: .infinity,alignment: .leading) // 최대 너비를 무한대로 설정하고 왼쪽 정렬
-            .padding() // 텍스트 주변에 패딩 추가
-            .background(Color.white) // 배경색을 흰색으로 설정
-            .overlay(
-                RoundedRectangle(cornerRadius: 10) // 모서리가 둥근 사각형
-                    .stroke(Color.gray, lineWidth: 1) // 회색 테두리
-                    .shadow(color: .cyan, radius: 30)
-            )
-            .padding() // 외부 패딩 추가
-    }
-    
-    func playToggle() {
-        if !isPlaying {
-            MusicPlayerManager.shared.playMusic(by: songInfo.id)
-            sharedDateManager.nowPlayingID = songInfo.id
-//            print("\(songInfo.id), \(sharedDateManager.nowPlayingID)")
-        } else {
-            MusicPlayerManager.shared.pauseMusic(by: songInfo.id)
-            sharedDateManager.nowPlayingID = ""
-//            print("\(songInfo.id), \(MusicPlayerManager.shared.nowPlayingID)")
-        }
-        isPlaying.toggle()
-    }
-    
-    @ViewBuilder
-    func menuButton() -> some View {
-        Menu {
-            Button(role: .destructive, action: showAlert) {
-                Label("삭제하기", systemImage: "trash")
-            }
-        } label: {
-            Image(systemName: "ellipsis")
-                .resizable()
-                .scaledToFit()
-                .frame(width: 15, height: 15)
-                .padding(10)
-                .foregroundColor(.gray)
-        }
-    }
-    
-    func showAlert() {
-        showingAlert = true
-    }
-    
-    func deleteRecord() {
-        modelContext.delete(recordArray[0])
-    }
-    
-}
-
-struct noRecordContentView: View {
-    @State private var showingModal = false
-    
-    var body: some View {
-        ZStack {
-            Rectangle()
-                .foregroundStyle(.cyan)
-                .ignoresSafeArea(.all, edges: .bottom)
-            
-            addRecordButton()
-            
-        }
-    }
-    
-    @ViewBuilder
-    func addRecordButton() -> some View {
-        Button(action: {
-            showingModal = true
-        }, label: {
-            
-            VStack(spacing: 10) {
-                Image(systemName: "plus")
-                    .font(.title)
-                
-                Text("기록 추가하기")
-                    .fontWeight(.medium)
-                    .font(.system(size: 20))
-            }
-        })
-        .foregroundStyle(.blue)
-        .frame(width: 150, height: 150)
-        .background(RoundedRectangle(cornerRadius: 20)
-            .foregroundStyle(.white)
-            .shadow(color: .gray, radius: 3, x: 0, y:0)
-        )
-        .sheet(isPresented: $showingModal) {
-            MusicSearchView(isPresented: $showingModal)
-        }
-    }
-}
-
-//#Preview {
-//    Home()
-//}
